requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.19.3
transformers>=4.30.0
safetensors>=0.3.1
accelerate>=0.20.3
xformers>=0.0.20
Pillow>=9.5.0
matplotlib>=3.7.1
gradio>=3.34.0
